+++ extension/sessionsifu@local/utils/fileUtils.js
@@ -26,6 +26,7 @@ export const config_path_base = GLib.build_filenamev([user_config, 'sessionsifu'
 export const sessions_path = GLib.build_filenamev([config_path_base, 'sessions']);
 export const history_path = GLib.build_filenamev([config_path_base, 'history']);
 export const history_limit = 5;
+export const recall_path = GLib.build_filenamev([config_path_base, 'recall']);
 export const sessions_backup_folder_name = 'backups';
 const sessions_backup_path = GLib.build_filenamev([sessions_path, sessions_backup_folder_name]);